+++ handlers/product-move-api/src/main/scala/com/gu/productmove/switchtype/RecurringContributionToSupporterPlus.scala
@@ -1,0 +1,318 @@
+package com.gu.productmove.endpoint.move
+
+import com.gu.newproduct.api.productcatalog.{Annual, BillingPeriod, Monthly}
+import com.gu.supporterdata.model.SupporterRatePlanItem
+import com.gu.productmove.endpoint.move.ProductMoveEndpointTypes.*
+import com.gu.productmove.GuStageLive.Stage
+import com.gu.productmove.framework.ZIOApiGatewayRequestHandler.TIO
+import com.gu.productmove.framework.{LambdaEndpoint, ZIOApiGatewayRequestHandler}
+import com.gu.productmove.refund.RefundInput
+import com.gu.productmove.salesforce.Salesforce.SalesforceRecordInput
+import com.gu.productmove.zuora.GetSubscription.RatePlanCharge
+import com.gu.productmove.zuora.rest.{ZuoraClientLive, ZuoraGet, ZuoraGetLive}
+import com.gu.util.config
+import com.gu.util.config.ZuoraEnvironment
+import com.gu.newproduct.api.productcatalog.ZuoraIds.{
+  ProductRatePlanId,
+  SupporterPlusZuoraIds,
+  ZuoraIds,
+  zuoraIdsForStage,
+}
+import com.gu.productmove.zuora.{
+  AddRatePlan,
+  GetAccount,
+  GetAccountLive,
+  GetSubscription,
+  GetSubscriptionLive,
+  RemoveRatePlan,
+  Subscribe,
+  SubscribeLive,
+  SubscriptionUpdate,
+  SubscriptionUpdateLive,
+  SubscriptionUpdatePreviewResponse,
+  SubscriptionUpdateRequest,
+  ZuoraCancel,
+  ZuoraCancelLive,
+  getContributionAmount,
+}
+import com.gu.productmove.{
+  AwsCredentialsLive,
+  AwsS3Live,
+  Dynamo,
+  DynamoLive,
+  EmailMessage,
+  EmailPayload,
+  EmailPayloadContactAttributes,
+  EmailPayloadProductSwitchAttributes,
+  GuStageLive,
+  SQS,
+  SQSLive,
+  SttpClientLive,
+}
+import sttp.tapir.*
+import sttp.tapir.EndpointIO.Example
+import sttp.tapir.Schema
+import sttp.tapir.json.zio.jsonBody
+import zio.{Clock, IO, URIO, ZIO}
+import zio.json.{DeriveJsonDecoder, DeriveJsonEncoder, JsonDecoder, JsonEncoder}
+import zio.ThreadLocalBridge.trace
+import com.gu.newproduct.api.productcatalog.PricesFromZuoraCatalog
+import com.gu.util.config.ZuoraEnvironment
+import com.gu.effects.GetFromS3
+import com.gu.newproduct.api.productcatalog.ZuoraIds.ZuoraIds
+
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+import com.gu.i18n.Currency
+import com.gu.productmove.zuora.model.SubscriptionName
+
+object RecurringContributionToSupporterPlus {
+  private def getSingleOrNotEligible[A](list: List[A], message: String): IO[String, A] =
+    list.length match {
+      case 1 => ZIO.succeed(list.head)
+      case _ => ZIO.fail(message)
+    }
+
+  def apply(
+      subscriptionName: SubscriptionName,
+      postData: ExpectedInput,
+  ): ZIO[GetSubscription with SubscriptionUpdate with GetAccount with SQS with Dynamo with Stage, String, OutputBody] =
+    (for {
+      _ <- ZIO.log("PostData: " + postData.toString)
+      subscription <- GetSubscription.get(subscriptionName).addLogMessage("GetSubscription")
+
+      currentRatePlan <- getSingleOrNotEligible(
+        subscription.ratePlans,
+        s"Subscription: ${subscriptionName.value} has more than one ratePlan",
+      )
+      ratePlanCharge <- getSingleOrNotEligible(
+        currentRatePlan.ratePlanCharges,
+        s"Subscription: ${subscriptionName.value} has more than one ratePlanCharge",
+      )
+      currency <- ZIO
+        .fromOption(ratePlanCharge.currencyObject)
+        .orElseFail(
+          s"Missing or unknown currency ${ratePlanCharge.currency} on rate plan charge in rate plan ${currentRatePlan.id} ",
+        )
+
+      result <-
+        if (postData.preview)
+          doPreview(
+            SubscriptionName(subscription.id),
+            postData.price,
+            ratePlanCharge.billingPeriod,
+            currency,
+            currentRatePlan.id,
+          )
+        else
+          doUpdate(
+            subscriptionName,
+            postData.price,
+            BigDecimal(ratePlanCharge.price),
+            ratePlanCharge.billingPeriod,
+            currency,
+            currentRatePlan,
+            subscription,
+            postData.csrUserId,
+            postData.caseId,
+          )
+    } yield result).fold(errorMessage => InternalServerError(errorMessage), success => success)
+
+    case class SupporterPlusRatePlanIds(
+        ratePlanId: String,
+        subscriptionRatePlanChargeId: String,
+        contributionRatePlanChargeId: Option[String],
+    )
+
+    def getSupporterPlusRatePlanIds(
+        stage: Stage,
+        billingPeriod: BillingPeriod,
+    ): Either[String, SupporterPlusRatePlanIds] = {
+      zuoraIdsForStage(config.Stage(stage.toString)).flatMap { zuoraIds =>
+        import zuoraIds.supporterPlusZuoraIds.{annual, annualV2, monthly, monthlyV2}
+
+        billingPeriod match {
+          case Monthly if SwitchToV2SupporterPlus =>
+            Right(
+              SupporterPlusRatePlanIds(
+                monthlyV2.productRatePlanId.value,
+                monthlyV2.productRatePlanChargeId.value,
+                Some(monthlyV2.contributionProductRatePlanChargeId.value),
+              ),
+            )
+          case Monthly =>
+            Right(
+              SupporterPlusRatePlanIds(
+                monthly.productRatePlanId.value,
+                monthly.productRatePlanChargeId.value,
+                None,
+              ),
+            )
+          case Annual if SwitchToV2SupporterPlus =>
+            Right(
+              SupporterPlusRatePlanIds(
+                annualV2.productRatePlanId.value,
+                annualV2.productRatePlanChargeId.value,
+                Some(annualV2.contributionProductRatePlanChargeId.value),
+              ),
+            )
+          case Annual =>
+            Right(
+              SupporterPlusRatePlanIds(
+                annual.productRatePlanId.value,
+                annual.productRatePlanChargeId.value,
+                None,
+              ),
+            )
+          case _ => Left(s"error when matching on billingPeriod $billingPeriod")
+        }
+      }
+    }
+
+    def getRatePlans(
+        billingPeriod: BillingPeriod,
+        currency: Currency,
+        ratePlanIdToRemove: String,
+        price: BigDecimal,
+    ): ZIO[Stage, String, (List[AddRatePlan], List[RemoveRatePlan])] =
+      for {
+        date <- Clock.currentDateTime.map(_.toLocalDate)
+        stage <- ZIO.service[Stage]
+        supporterPlusRatePlanIds <- ZIO.fromEither(getSupporterPlusRatePlanIds(stage, billingPeriod))
+        overrideAmount <- getContributionAmount(stage, price, currency, billingPeriod)
+        chargeOverride = ChargeOverrides(
+          price = Some(overrideAmount),
+          productRatePlanChargeId = supporterPlusRatePlanIds.contributionRatePlanChargeId.getOrElse(
+            supporterPlusRatePlanIds.subscriptionRatePlanChargeId,
+          ),
+        )
+        addRatePlan = AddRatePlan(date, supporterPlusRatePlanIds.ratePlanId, chargeOverrides = List(chargeOverride))
+        removeRatePlan = RemoveRatePlan(date, ratePlanIdToRemove)
+      } yield (List(addRatePlan), List(removeRatePlan))
+
+    def doPreview(
+        subscriptionName: SubscriptionName,
+        price: BigDecimal,
+        billingPeriod: BillingPeriod,
+        currency: Currency,
+        currentRatePlanId: String,
+    ): ZIO[SubscriptionUpdate with Stage, String, OutputBody] = for {
+      _ <- ZIO.log("Fetching Preview from Zuora")
+
+      today <- Clock.currentDateTime.map(_.toLocalDate)
+
+      updateRequestBody <- getRatePlans(billingPeriod, currency, currentRatePlanId, price).map {
+        case (addRatePlan, removeRatePlan) =>
+          SubscriptionUpdateRequest(add = addRatePlan, remove = removeRatePlan, targetDate = today.plusMonths(13))
+      }
+
+      response <- SubscriptionUpdate
+        .update[SubscriptionUpdatePreviewResponse](subscriptionName, updateRequestBody)
+        .addLogMessage("SubscriptionUpdate")
+
+      stage <- ZIO.service[Stage]
+      supporterPlusRatePlanIds <- ZIO.fromEither(getSupporterPlusRatePlanIds(stage, billingPeriod))
+      previewResult <- BuildPreviewResult.getPreviewResult(response.invoice, supporterPlusRatePlanIds)
+    } yield previewResult
+
+    def doUpdate(
+        subscriptionName: SubscriptionName,
+        price: BigDecimal,
+        previousAmount: BigDecimal,
+        billingPeriod: BillingPeriod,
+        currency: Currency,
+        currentRatePlan: GetSubscription.RatePlan,
+        subscription: GetSubscription.GetSubscriptionResponse,
+        csrUserId: Option[String],
+        caseId: Option[String],
+    ): ZIO[GetAccount with SubscriptionUpdate with SQS with Stage, String, OutputBody] = for {
+      _ <- ZIO.log("Performing product move update")
+      stage <- ZIO.service[Stage]
+      account <- GetAccount.get(subscription.accountNumber).addLogMessage("GetAccount")
+
+      identityId <- ZIO
+        .fromOption(account.basicInfo.IdentityId__c)
+        .orElseFail(s"identityId is null for subscription name ${subscriptionName.value}")
+
+      updateRequestBody <- getRatePlans(billingPeriod, currency, ratePlanIdToRemove, price).map {
+        case (addRatePlan, removeRatePlan) =>
+          SubscriptionUpdatePreviewRequest(add = addRatePlan, remove = removeRatePlan, targetDate = targetDate)
+      }
+
+      updateResponse <- SubscriptionUpdate
+        .update[SubscriptionUpdateResponse](subscriptionName, updateRequestBody)
+        .addLogMessage("SubscriptionUpdate")
+
+      todaysDate <- Clock.currentDateTime.map(_.toLocalDate)
+      billingPeriodValue <- billingPeriod.value
+
+      paidAmount = updateResponse.paidAmount.getOrElse(BigDecimal(0))
+
+      emailFuture <- SQS
+        .sendEmail(
+          message = EmailMessage(
+            EmailPayload(
+              Address = Some(account.billToContact.workEmail),
+              ContactAttributes = EmailPayloadContactAttributes(
+                SubscriberAttributes = EmailPayloadProductSwitchAttributes(
+                  first_name = account.billToContact.firstName,
+                  last_name = account.billToContact.lastName,
+                  currency = account.basicInfo.currency.symbol,
+                  price = price.setScale(2, BigDecimal.RoundingMode.FLOOR).toString,
+                  first_payment_amount = paidAmount.setScale(2, BigDecimal.RoundingMode.FLOOR).toString,
+                  date_of_first_payment = todaysDate.format(DateTimeFormatter.ofPattern("d MMMM uuuu")),
+                  payment_frequency = billingPeriodValue + "ly",
+                  subscription_id = subscriptionName.value,
+                ),
+              ),
+            ),
+            "SV_RCtoSP_Switch",
+            account.basicInfo.sfContactId__c,
+            Some(identityId),
+          ),
+        )
+        .fork
+
+      salesforceTrackingFuture <- SQS
+        .queueSalesforceTracking(
+          SalesforceRecordInput(
+            subscriptionName.value,
+            previousAmount,
+            price,
+            currentRatePlan.productName,
+            currentRatePlan.ratePlanName,
+            "Supporter Plus",
+            todaysDate,
+            todaysDate,
+            paidAmount,
+            csrUserId,
+            caseId,
+          ),
+        )
+        .fork
+
+      supporterPlusRatePlanIds <- ZIO.fromEither(getSupporterPlusRatePlanIds(stage, billingPeriod))
+      amendSupporterProductDynamoTableFuture <- Dynamo
+        .writeItem(
+          SupporterRatePlanItem(
+            subscriptionName.value,
+            identityId = identityId,
+            gifteeIdentityId = None,
+            productRatePlanId = supporterPlusRatePlanIds.ratePlanId,
+            productRatePlanName = "product-move-api added Supporter Plus Monthly",
+            termEndDate = todaysDate.plusDays(7),
+            contractEffectiveDate = todaysDate,
+            contributionAmount = None,
+          ),
+        )
+        .fork
+
+      requests = emailFuture
+        .zip(salesforceTrackingFuture)
+        .zip(amendSupporterProductDynamoTableFuture)
+
+      _ <- requests.join
+
+    } yield Success("Product move completed successfully")
+
+}
